attempt cleanup in text_18118.c

+++ include/text_24E10.h
@@ -11,7 +11,7 @@ extern u8 D_801CEFA4;
 void INIT_TEXT_TO_DISPLAY(void);
 u8 deter_num_let(u8 param_1);
 void calc_num_let_spr(u8 param_1, u8 *param_2);
-s32 calc_largmax_text(u8 *param_1, s32 param_2, s16 space_width, s16 param_4, u8 param_5);
+s16 calc_largmax_text(u8 *param_1, s32 param_2, s16 space_width, s16 param_4, u8 param_5);
 void INIT_TXT_BOX(TextToDisplay *out_txt);
 
 #endif
